feat: slide changes EXLM-3374
