+++ blocks/slides/slider-state.js
